# Remove instance of DateTimeOffset in Span.Finish